moved auth route function to its own source file
