+++ frontend/src/main.js
@@ -8,8 +8,6 @@ import { aliases, mdi } from "vuetify/iconsets/mdi"
 import * as components from "vuetify/components"
 import * as directives from "vuetify/directives"
 
-import Cookies from "universal-cookie";
-
 /* Page Components */
 const HomePage = () => import("@/components/pages/HomePage.vue");
 const AboutPage = () => import("@/components/pages/AboutPage.vue");
@@ -18,26 +16,7 @@ const DashboardPage = () => import("@/components/pages/DashboardPage.vue");
 const LoginPage = () => import("@/components/pages/LoginPage.vue");
 const RegisterPage = () => import("@/components/pages/RegisterPage.vue");
 
-async function AuthRoute() {
-  let is_authenticated = false;
-
-  const cookies = new Cookies();
-  const token = cookies.get("token");
-
-  const body = { token: token };
-
-  const response = await fetch("http://localhost:8000/api/get-decoded-token", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(body)
-  });
-  const response_json = await response.json();
-  is_authenticated = response_json["auth"];
-
-  return is_authenticated;
-}
+import AuthRoute from "@/components/AuthRoute";
 
 const routes = [
   {
